requirements.txt
openai
beautifulsoup4
streamlit
pdfkit
python-dotenv
python-jose[cryptography]
fastapi
python-multipart
uvicorn